requirements.txt
# Python packages required for classify_picamera.py
numpy
Pillow
flask